--- a/dev_log/2.2.8/crayon_202203172040.yaml
+++ b/dev_log/2.2.8/crayon_202203172040.yaml
@@ -1,0 +1,3 @@
+---
+optimization:
+  - "插件订阅变更计算优化 (closed #599)"