## 运行环境
* Mac 8c16g

## 效果

**73s -> 15.2s**

## 优化前
```
Total time: 52.9118 s
apps/backend/subscription/steps/plugin.py
Function: make_instances_migrate_actions at line 480
check_config_change (92.6%)
---
Total time: 14.5445 s
apps/backend/subscription/tasks.py
Function: create_task at line 235
89.6%    host_id__bk_obj_sub_map = models.Subscription.get_host_id__bk_obj_sub_map(bk_host_ids, step_id)
---
Total time: 4.99353 s
apps/backend/subscription/tasks.py
Function: run_subscription_task_and_create_instance at line 445
tools.get_instances_by_scope(scope) （99.5%）
```

## 优化后

```
Total time: 4.59554 s
apps/backend/subscription/steps/plugin.py
Function: make_instances_migrate_actions at line 555
Line 689 self.handle_config_change_instances (65%)
---
Total time: 2.32126 s
apps/backend/subscription/tasks.py
Function: create_task at line 235
8.1% host_id__bk_obj_sub_map = models.Subscription.get_host_id__bk_obj_sub_map(bk_host_ids, step_id)
---
Total time: 6.46358 s
apps/backend/subscription/tasks.py
Function: run_subscription_task_and_create_instance at line 445
tools.get_instances_by_scope(scope) （98%）
```